lib: obstacles: move class Circle to its own file
