+++ examples/lidar_obstacles/lidar_obstacles.cpp
@@ -11,7 +11,7 @@
 // Application includes
 #include "common_defs.h"
 #include "lidar_utils.hpp"
-#include "obstacles/obstacles.hpp"
+#include "obstacles/Circle.hpp"
 #include "obstacles/List.hpp"
 #include "cogip_defs/Coords.hpp"
 #include "cogip_defs/Pose.hpp"
