+++ platforms/cortex/lidar_obstacles.cpp
@@ -11,7 +11,7 @@
 #include "lds01.h"
 
 // Application includes
-#include "obstacles/obstacles.hpp"
+#include "obstacles/Circle.hpp"
 #include "platform.hpp"
 
 // Number of angles without obstacle to ignore between two angles with obstacle
